refactor: 상수 줄바꿈, 메서드명 변경, 패키지 이동
