+++ src/main/java/roomescape/controller/dto/reservation/ReservationMemberRequest.java
@@ -4,7 +4,7 @@ import jakarta.validation.constraints.NotNull;
 
 import java.time.LocalDate;
 
-public record ReservationRequest(
+public record ReservationMemberRequest(
         @NotNull LocalDate date,
         long timeId,
         long themeId
